Time: 1 ms (9.76%), Space: 47.7 MB (71.53%) - LeetHub

+++ 0050-powx-n/0050-powx-n.java
@@ -1,6 +1,6 @@
 class Solution {
     public double myPow(double x, int n) {
-        long N = n;
+        int N = n;
         return power(x, N);
     }
 
@@ -9,7 +9,7 @@ class Solution {
         if (n == 1) return x;
 
         if (n < 0) {
-            return 1 / power(x, -n);
+            return 1.0 / power(x, -n);
         } else if (n % 2 == 0) {
             return power(x * x, n / 2);
         } else {
